ntdb: switch between secrets.tdb and secrets.ntdb depending on 'use ntdb'
Since we open with dbwrap, it auto-converts old tdbs (which it will
rename to secrets.tdb.bak once it's done).


Autobuild-User(master): Rusty Russell <rusty@rustcorp.com.au>
Autobuild-Date(master): Wed Feb 20 07:09:19 CET 2013 on sn-devel-104

+++ source4/scripting/python/samba/tests/provision.py
@@ -55,8 +55,9 @@ class ProvisionTestCase(samba.tests.TestCaseInTempDir):
 
     def test_setup_secretsdb(self):
         path = os.path.join(self.tempdir, "secrets.ldb")
-        secrets_tdb_path = os.path.join(self.tempdir, "secrets.tdb")
         paths = ProvisionPaths()
+        secrets_tdb_path = os.path.join(self.tempdir, "secrets.tdb")
+        secrets_ntdb_path = os.path.join(self.tempdir, "secrets.ntdb")
         paths.secrets = path
         paths.private_dir = os.path.dirname(path)
         paths.keytab = "no.keytab"
@@ -68,8 +69,10 @@ class ProvisionTestCase(samba.tests.TestCaseInTempDir):
         finally:
             del ldb
             os.unlink(path)
-            os.unlink(secrets_tdb_path)
-
+            if os.path.exists(secrets_tdb_path):
+                os.unlink(secrets_tdb_path)
+            if os.path.exists(secrets_ntdb_path):
+                os.unlink(secrets_ntdb_path)
 
 class FindNssTests(TestCase):
     """Test findnss() function."""
